Instruction and data pointers can now be modified at any time.
Instruction pointer cannot be set to less than zero or more than the number
of instructions in current code. Prints out warning for users otherwise.

Data pointer can be set to any value, but will flash red if pointing outside
of memory.

Cleaned code to include a display data pointer that eliminates need to
query browser for it's address.