Many small changes made to user interface.
Instruction pointer highlights when changed.

Clicking on Instruction and Data pointer automatically selects the contents of the input.  This is a convenience to the user.

Created button to automatically copy and paste sample code from the examples in the documentation.

Added alternate text for each button in the main interface. Allows users with screen readers to use the tool.

Instruction Reference has a better look and feel.